new way of adding/editing activities, added activity description, refactored code

+++ lib/screens/profile/screens/manage_profile/screens/delete_profile.dart
@@ -2,14 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:iconsax/iconsax.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:sc_app/providers/auth.dart';
-import 'package:fluentui_system_icons/fluentui_system_icons.dart';
-import 'package:sc_app/helpers/show.dart';
+import 'package:sc_app/helpers/helpers.dart';
 import 'package:sc_app/utils/enums.dart';
 import 'package:sc_app/widgets/alerts.dart';
 import 'package:sc_app/widgets/buttons.dart';
 import 'package:sc_app/widgets/bullet_list.dart';
 import 'package:sc_app/widgets/loading.dart';
-import 'package:sc_app/widgets/textfield_label.dart';
+import 'package:sc_app/widgets/input_field_label.dart';
 
 class DeleteProfileScreen extends ConsumerStatefulWidget {
   const DeleteProfileScreen({super.key});
@@ -75,18 +74,26 @@ class _DeleteProfileScreenState extends ConsumerState<DeleteProfileScreen> {
           const SizedBox(height: 4),
           BulletList(
             texts: [
-              Text(
-                'All your synced data/activities will be permanently deleted.',
-                style: TextStyle(color: Theme.of(context).colorScheme.error),
+              RichText(
+                text: TextSpan(
+                  style: TextStyle(color: Theme.of(context).colorScheme.error),
+                  children: const [
+                    TextSpan(text: 'All your synced data/activities will be'),
+                    TextSpan(
+                      text: 'permanently deleted.',
+                      style: TextStyle(fontWeight: FontWeight.bold),
+                    ),
+                  ],
+                ),
               ),
               RichText(
                 text: TextSpan(
-                  text: 'Your profile will be permanently deleted, and ',
                   style: TextStyle(color: Theme.of(context).colorScheme.error),
                   children: const [
+                    TextSpan(text: 'Your profile will be'),
                     TextSpan(
-                      text: 'you won\'t be able to revert this action.',
-                      style: TextStyle(fontWeight: FontWeight.w600),
+                      text: 'permanently deleted,',
+                      style: TextStyle(fontWeight: FontWeight.bold),
                     ),
                   ],
                 ),
@@ -94,8 +101,8 @@ class _DeleteProfileScreenState extends ConsumerState<DeleteProfileScreen> {
             ],
           ),
           const SizedBox(height: 32),
-          const TextFieldLabel(
-            text: 'Enter Your Password',
+          const InputFieldLabel(
+            label: 'Enter Your Password',
             bottomPadding: 8,
           ),
           Form(
@@ -141,33 +148,22 @@ class _DeleteProfileScreenState extends ConsumerState<DeleteProfileScreen> {
             onPressed: () {
               setState(() => errorMessage = null);
               if (_formKey.currentState!.validate()) {
-                Show.modal(
-                  context,
-                  modal: Alert(
-                    title: Text(
-                      'Delete Profile',
+                showDialog(
+                  context: context,
+                  builder: (context) => ConfirmationAlert(
+                    title: const Text('Delete Profile?'),
+                    content: Text(
+                      'This is final! You won\'t be able to revert this action.',
                       style: TextStyle(
                         color: Theme.of(context).colorScheme.error,
                       ),
                     ),
-                    titleIcon: Icon(
-                      FluentIcons.warning_24_filled,
-                      color: Theme.of(context).colorScheme.error,
-                    ),
-                    content: const Text(
-                      'Are you sure you want to delete this profile.'
-                      '\nYou won\'t be able to revert this action.',
-                    ),
-                    actionName: 'Delete',
                     action: () => _deleteProfile(
-                      onDone: () {
-                        Navigator.pop(context);
-                        Show.snackBar(
-                          context,
-                          text: 'Profile was successfully deleted.',
-                          snackBarIcon: SnackBarIcon.done,
-                        );
-                      },
+                      onDone: () => Helpers.showSnackBar(
+                        context,
+                        text: 'Profile was successfully deleted.',
+                        snackBarIcon: SnackBarIcon.done,
+                      ),
                     ),
                   ),
                 );
